Fix gitignore and some Constants type

+++ src/Constants.hpp
@@ -6,7 +6,7 @@ namespace Constants
 {
     // Maximum number of objects to group in a single render group
     // This also defines the size of the uniform buffers in standard shaders
-    const unsigned long long maxObjectsToGroup = 8192;
+    const unsigned long maxObjectsToGroup = 8192;
 }
 
 #endif
